Remove layout conversion from #blocked to #mma
In the streamk kernel we have the following ir
```mlir
%val = tt.load %ptr tensor<#blocked>
%0 = ttg.convert_layout %val tensor<#blocked> --> tensor<#mma>
%1 = arith.addf %0, %acc : tensor<#mma>
```
in which the computed acc needs to be accumulated with saved partial
sum from other pid. %val and %acc both are very large, and take a lot
of vgprs. Ideally, we want to interleave the loaded value and
accumulation of the partial sum. However, the convert_layout in
between forces it to finish load all values before starting the
accumulation.

By adding a remove_layout_conversion pass at the end of the pass
pipeline, this convert_layout is removed.