[LAYOUTS] Generic stmatrix lowering (#6609)
We use `divideLeft` to lower a generic `local_store` using `stmatrix`
whenever possible.

We implement `ColumnAction` as a helper that allows us to permute
the bases and values and remove broadcasting in a generic way.

The current codegen for `stmatrix` has some limitations, so we just
return early in those cases. We'll fix those in a future PR.