[AMD] Reject async global to local ops for less than 4 bytes (#6734)
Direct-to-lds loads for smaller than 4 bytes extend their loaded values
to 4 bytes before storing to shared memory. Adjusted lit tests from f16
to f32 so they produce >= 4 byte loads.

This does not affect what we have because the streampipeliner will
pipeline those loads via registers, but we should ensure we do not
produce incorrect assembly if that changes in the future.