[AMD] Add an environment variable to disable buffer atomics (#7311)
Adds an environment variable, `AMDGCN_DISABLE_BUFFER_ATOMICS`
that can be set to disable Buffer atomics on the MI300. We have seen
significant wins from BufferOps in general, but we haven't benchmarked
the impact of atomics at Meta and we have seen some issue internally.

We are working to resolve these issues/will open reproducers upstream,
but with BufferOps being the default we would be much happier if we
could allow a production workload to opt to accept BufferOps but not
atomics.