Force swizzledSharedLayout for all users of load
In hstu bwd kernel, the load has two users
1. regular cvt from #blocked to #dotOp
2. special cvt from #blocked to #linear

The second cvt is doing the actually transpose operation so that the
following transOp is a no-op.
Currently we only expect #linear for scaled fp8 operands and a special
LDS layout is chosen.
This commit simply forces every user to use the swizzled LDS layout to
enable the loads to be pipelined