Merge pr 7311 and 7939  to the test branch (#885)
* [AMD] Add an environment variable to disable buffer atomics (#7311)

Adds an environment variable, `AMDGCN_DISABLE_BUFFER_ATOMICS`
that can be set to disable Buffer atomics on the MI300. We have seen
significant wins from BufferOps in general, but we haven't benchmarked
the impact of atomics at Meta and we have seen some issue internally.

We are working to resolve these issues/will open reproducers upstream,
but with BufferOps being the default we would be much happier if we
could allow a production workload to opt to accept BufferOps but not
atomics.

* [AMD] Accumulate into offset instead of pointer for jit specialized tensor (#7939)

We canonicalize pointer pointing to small-tensor in different way. For
example, given input like this:
   %p1 = tt.addptr %p0, %ofst
    ...
   %p2 = tt.addptr %p1, %ofst2

it will be canonicalized into following:
   %p1 = tt.addptr %p0, %ofst
    ...
   %p2 = tt.addptr %p0, (%ofst2 + %ofst)
   
The rationale is two-fold:
1) is to fix bug like the one reported in the issue-830
2) to reveal the information to buffer-op optimization. buffer-op pass
will easily spot the a global memory op whose base pointer is pointing
to a small-tensor and hence can safely convert them into buffer-op.

---------

Co-authored-by: Nick Riasanovsky <njriasan@meta.com>
Co-authored-by: yangshuxin <shuxin.yang@amd.com>