[AMD][BACKEND] Adjust `UpdateAsyncWaitCount` to work without tokens  (#217)
Implements the proposal in https://github.com/ROCm/triton-internal/issues/1318. For Gluon instead of following the tokens, which are difficult to loop carry in gluon, we define the waitCnt as number of commit groups ops in Gluon. The UpdateAsyncWaitCnt pass will then walk the IR backwards and computes the minimum number of created async intrinsic by waitCnt outstanding commit groups.

Current limitations are:

We treat dynamic and static loops the same. This means we could get conservative loads if the prologue prefetch loop is not unrolled. For our Gluon kernels they were always unrolled. I can add support in a follow up PR since this one is already quite large.
scf.ExecuteRegion is not handled to keep the PR minimal. But adding support for it is trivial when the new PingPong lands
Warp specialization is not handled. We will probably use LDS/Named barriers so this might not be needed? Adding it should also be not a problem in case we need it.
Note that for TDM each tensor_load/store in TGGIR will create exactly one intrinsic so we can use the count from Gluon directly.